% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-annotate.R
\name{annotate}
\alias{annotate}
\title{This is a title for a function}
\usage{
annotate()
}
\arguments{
\item{parameter}{Here's a parameter let's describe it here}
}
\description{
This is a function here's where we describe what it does
}
\examples{
\dontrun{

}
}
