% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-crispr-calc.R
\name{calc_crispr}
\alias{calc_crispr}
\title{Calculate CRISPR scores}
\usage{
calc_crispr(.data = NULL, gimap_dataset, normalized = TRUE)
}
\arguments{
\item{.data}{Data can be piped in with tidyverse pipes from function to function. But the data must still be a gimap_dataset}

\item{gimap_dataset}{A special dataset structure that is setup using the `setup_data()` function.}

\item{normalized}{Default is TRUE meaning that we should expect to look for normalized data in the gimap_dataset.}
}
\description{
This calculates the log fold change for a gimap dataset based on the annotation and metadata provided.
}
\examples{
\dontrun{

gimap_dataset <- get_example_data("gimap")

# Highly recommended but not required
run_qc(gimap_dataset)

gimap_dataset <- gimap_dataset \%>\%
  gimap_filter() \%>\%
  gimap_annotate(cell_line = "HELA") \%>\%
  gimap_normalize(
    timepoints = "day",
    replicates = "rep"
  ) \%>\%
  calc_crispr()

# To see results
gimap_dataset$crispr_score
}
}
