% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foldchange.R
\name{calc_fc}
\alias{calc_fc}
\title{This is a title for a function}
\usage{
calc_fc()
}
\arguments{
\item{parameter}{Here's a parameter let's describe it here}
}
\description{
This is a function here's where we describe what it does
}
\examples{
\dontrun{

}
}
