% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-calculate_gi.R
\name{calc_gi}
\alias{calc_gi}
\title{Calculate Genetic Interaction scores}
\usage{
calc_gi(.data = NULL, gimap_dataset)
}
\arguments{
\item{.data}{Data can be piped in with tidyverse pipes from function to function. But the data must still be a gimap_dataset}

\item{gimap_dataset}{A special dataset structure that is setup using the `setup_data()` function.}
}
\description{
Create results table that has CRISPR scores, Wilcoxon rank-sum test and t tests.
}
\examples{
{
  gimap_dataset <- get_example_data("gimap")

  # Highly recommended but not required
  run_qc(gimap_dataset)

  gimap_dataset <- gimap_dataset \%>\%
    gimap_filter() \%>\%
    gimap_annotate(cell_line = "HELA") \%>\%
    gimap_normalize(
      timepoints = "day",
      replicates = "rep"
    ) \%>\%
    calc_crispr() \%>\%
    calc_gi()

  saveRDS(gimap_dataset, "gimap_dataset_final.RDS")
}
}
