% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-foldchange.R
\name{calc_lfc}
\alias{calc_lfc}
\title{Calculate log fold change for a}
\usage{
calc_lfc(.data = NULL, gimap_dataset, replicates = NULL, timepoints = NULL)
}
\arguments{
\item{.data}{Data can be piped in with %>% or |> from function to function. But the data must still be a gimap_dataset}

\item{gimap_dataset}{A special dataset structure that is setup using the `setup_data()` function.}

\item{replicates}{Specifies the column name of the metadata set up in `$metadata$sample_metadata` that has a factor that represents the replicates.}

\item{timepoints}{Specifies the column name of the metadata set up in `$metadata$sample_metadata` that has a factor that represents the timepoints.}
}
\description{
This calculates the log fold change for a gimap dataset based on the annotation and metadata provided.
}
\examples{
\dontrun{

gimap_dataset <- get_example_data("gimap")

# Highly recommended but not required
run_qc(gimap_dataset)

gimap_dataset <- gimap_dataset \%>\%
  gimap_filter() \%>\%
  gimap_annotate() \%>\%
  calc_lfc()

# To see results
gimap_dataset$log_fc
}
}
