% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{config_folder}
\alias{config_folder}
\title{Get file path to config folder}
\usage{
config_folder()
}
\value{
Returns the file path to folder where the example data is stored
}
\description{
Get file path to config folder
}
