% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{example_data}
\alias{example_data}
\title{This is a title for a function}
\usage{
example_data()
}
\arguments{
\item{parameter}{Here's a parameter let's describe it here}
}
\description{
This is a function here's where we describe what it does
}
\examples{
\dontrun{

pg_data <- example_data()
}
}
