% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment.R
\name{fastq_to_bam}
\alias{fastq_to_bam}
\title{Align fastq files to create bam files}
\usage{
fastq_to_bam(
  fastq_dir,
  for_index,
  rev_index,
  sample_names,
  output_dir = "aligned_bam",
  save_logs = FALSE,
  time = TRUE,
  overwrite = TRUE
)
}
\arguments{
\item{fastq_dir}{a file path to a directory where the bam files associated with the sample names will be stored.}

\item{for_index}{A file path to the index that was made with `bowtie2_build()` for the forward reference file for the paired guide constructs}

\item{rev_index}{A file path to the index that was made with `bowtie2_build()` for the reverse reference file for the paired guide constructs}

\item{sample_names}{a character vector that indicates the name of the samples as they are listed in the file names.}

\item{output_dir}{A file path that designates where you'd like the aligned files saved. By default this file path is "aligned_bams"}

\item{save_logs}{TRUE/FALSE you want the logs? They will be saved as the sample name}

\item{time}{TRUE/FALSE you want the duration this takes to run printed out}

\item{overwrite}{TRUE/FALSE should existing bam files of the same name be overwritten?}
}
\description{
This function takes a forward and reverse bam files for a group of samples and returns the counts and stats.
}
\examples{
\dontrun{

#################### Build the indices first #################

# Declare an input file path for FORWARD
for_ref <- file.path(config_folder(), "ref", "paralog_pgRNA1.fa")
# Declare ann output file path for FORWARD
for_index_file_path <- file.path(config_folder(), "pgPEN_index_for")
# Declare an input file path for REVERSE
rev_ref <- file.path(config_folder(), "ref", "paralog_pgRNA2.fa")
# Declare a output file path for REVERSE
rev_index_file_path <- file.path(config_folder(), "pgPEN_index_rev")

# Build the FORWARD reference
bowtie2_build(
  references = for_ref,
  bt2Index = for_index_file_path,
  overwrite = TRUE
)

# Build the REVERSE reference
bowtie2_build(
  references = rev_ref,
  bt2Index = rev_index_file_path,
  overwrite = TRUE
)

# This will be the directory that contains all the file path
fastq_dir <- file.path(example_data_folder(), "fastqs", "fastq_demuxed")

# These MUST be names that are listed in the bam file name itself.
# There needs to be exactly 2 bam files per sample
sample_names <- c("sample1", "sample2", "sample3")

fastq_to_bam(
  fastq_dir = fastq_dir,
  for_index = for_index_file_path,
  rev_index = rev_index_file_path,
  sample_names = sample_names,
  output_dir = file.path(example_data_folder(), "aligned_bam"),
  time = TRUE,
  overwrite = TRUE
)
}
}
