% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_example_data}
\alias{get_example_data}
\title{Returns example data for package}
\usage{
get_example_data(which_data)
}
\arguments{
\item{which_data}{options are "count" or "meta"; specifies which example dataset should be returned}
}
\description{
This function loads and returns example data for the packagae. Which dataset is returned must be specified
}
\examples{
\dontrun{

gimap_dataset <- get_example_data("gimap")
}
}
