% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gimap-package.R
\docType{package}
\name{gimap-package}
\alias{gimap}
\alias{gimap-package}
\title{gimap: Calculate Genetic Interactions for CRISPR Targets}
\description{
This package allows a pipeline to be built for analyzing genetic interactions of paired guide CRISPR screens. It is based off of original research from the Alice Berger lab at Fred Hutchinson Cancer Center
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/FredHutch/gimap}
  \item Report bugs at \url{https://github.com/FredHutch/gimap/issues}
}

}
\author{
\strong{Maintainer}: Candace Savonen \email{cansav09@gmail.com, csavonen@fredhutch.org}

Authors:
\itemize{
  \item Howard Baek \email{howardbaek@fredhutch.org, howardbaek.fh@gmail.com}
  \item Kate Isaac \email{kisaac@fredhutch.org}
}

}
\keyword{internal}
