% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-annotate.R
\name{gimap_annotate}
\alias{gimap_annotate}
\title{Annotate gimap data}
\usage{
gimap_annotate(
  .data = NULL,
  gimap_dataset,
  cell_line,
  control_genes = NULL,
  cn_annotate = TRUE,
  annotation_file = NULL
)
}
\arguments{
\item{.data}{Data can be piped in with tidyverse pipes from function to function. But the data must still be a gimap_dataset}

\item{gimap_dataset}{A special dataset structure that is setup using the `setup_data()` function.}

\item{cell_line}{which cell line are you using? (e.g., HELA, PC9, etc.). Required argument}

\item{control_genes}{A vector of gene symbols (e.g. AAMP) that should be labeled as control genes. These will be used for log fold change calculations. If no list is given then DepMap Public 23Q4 Achilles_common_essentials.csv is used https://depmap.org/portal/download/all/}

\item{cn_annotate}{TRUE or FALSE you'd also like to have Copy number annotation from DepMap. These data are optional}

\item{annotation_file}{If no file is given, will attempt to use the design file from https://media.addgene.org/cms/filer_public/a9/9a/a99a9328-324b-42ff-8ccc-30c544b899e4/pgrna_library.xlsx}
}
\description{
In this function, a `gimap_dataset` is annotated as far as which genes should be used as controls.
}
\examples{
\dontrun{

gimap_dataset <- get_example_data("gimap")

# Highly recommended but not required
run_qc(gimap_dataset)

gimap_dataset <- gimap_dataset \%>\%
  gimap_filter() \%>\%
  gimap_annotate(cell_line = "HELA")

# To see anotations
gimap_dataset$annotation
}
}
