% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-filter.R
\name{gimap_filter}
\alias{gimap_filter}
\title{A function to run filtering}
\usage{
gimap_filter(.data = NULL, gimap_dataset, filter_type = "both")
}
\arguments{
\item{.data}{Data can be piped in with %>% or |> from function to function. But the data must still be a gimap_dataset}

\item{gimap_dataset}{A special dataset structure that is setup using the `setup_data()` function.}

\item{filter_type}{Can be one of the following: `zero_count_only`, `low_plasmid_cpm_only` or `rep_variation`, `zero_in_last_time_point` or a vector that includes multiple of these filters.
You should decide on the appropriate filter based on the results of your QC report.}
}
\value{
a filtered version of the gimap_dataset returned in the $filtered_data section
}
\description{
This is a function here's where we describe what it does
}
\examples{
\dontrun{

gimap_dataset <- get_example_data("gimap")

# Highly recommended but not required
run_qc(gimap_dataset)


gimap_dataset <- gimap_filter(gimap_dataset)

# To see filtered data
gimap_dataset$filtered_data

}

}
