% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-calculate_gi.R
\name{gimap_stats}
\alias{gimap_stats}
\title{Do tests for each sample}
\usage{
gimap_stats()
}
\arguments{
\item{.data}{Data can be piped in with %>% or |> from function to function. But the data must still be a gimap_dataset}

\item{gimap_dataset}{A special dataset structure that is setup using the `setup_data()` function.}

\item{test}{options include 'wilcoxon' and 't-test'. By default, both will be run.}

\item{overwrite}{default is FALSE; whether to overwrite the QC Report file}

\item{output_file}{default is `GI_Results`; name of the output GI results file}
}
\description{
Create results table that has CRISPR scores, Wilcoxon rank-sum test and t tests.
}
\examples{
\dontrun{

gimap_dataset <- get_example_data("gimap")

# Highly recommended but not required
run_qc(gimap_dataset)

gimap_dataset <- gimap_dataset \%>\%
  gimap_filter() \%>\%
  gimap_annotate() \%>\%
  gimap_normalize(
    timepoints = "day",
    replicates = "rep") \%>\%
  calc_crispr()



}
}
