% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-qc.R
\name{qc_constructs_countzero_bar}
\alias{qc_constructs_countzero_bar}
\title{Create a bar graph that shows the number of replicates with a zero count for pgRNA constructs flagged by the zero count filter}
\usage{
qc_constructs_countzero_bar(gimap_dataset, wide_ar = 0.75)
}
\arguments{
\item{gimap_dataset}{The special gimap_dataset from the `setup_data` function which contains the transformed data}

\item{wide_ar}{aspect ratio, default is 0.75}
}
\value{
a ggplot barplot
}
\description{
A short description...
}
\examples{
\dontrun{
gimap_dataset <- get_example_data("gimap")
qc_constructs_countzero_bar(gimap_dataset)
}

}
