% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-filter.R
\name{qc_filter_zerocounts}
\alias{qc_filter_zerocounts}
\title{Create a filter for pgRNAs which have a raw count of 0 for any sample/time point}
\usage{
qc_filter_zerocounts(gimap_dataset)
}
\arguments{
\item{gimap_dataset}{The special gimap_dataset from the `setup_data` function which contains the transformed data}
}
\value{
a named list with the filter `filter` specifying which pgRNA have a count zero for at least one sample/time point and a report df `reportdf` for the number and percent of pgRNA which have a count zero for at least one sample/time point
}
\description{
This function flags and reports which and how many pgRNAs have a raw count of 0 for any sample/time point
}
\examples{
\dontrun{

}

}
