% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-qc.R
\name{qc_plasmid_histogram}
\alias{qc_plasmid_histogram}
\title{Create a histogram with plasmid log2 CPM values and ascertain a cutoff for low values}
\usage{
qc_plasmid_histogram(gimap_dataset, cutoff = NULL, wide_ar = 0.75)
}
\arguments{
\item{gimap_dataset}{The special gimap_dataset from the `setup_data` function which contains the transformed data}

\item{cutoff}{default is NULL, the cutoff for low log2 CPM values for the plasmid time period}

\item{wide_ar}{aspect ratio, default is 0.75}
}
\value{
a named list
}
\description{
Find the distribution of plasmid (day0 data) pgRNA log2 CPM values, and ascertain a cutoff or filter for low log2 CPM values.
Assumes the first column of the dataset is the day0 data; do I need a better
method to tell, especially if there are reps?
}
