% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-setup_data.R
\name{setup_data}
\alias{setup_data}
\title{Making a new gimap dataset}
\usage{
setup_data(counts = NULL, pg_metadata = NULL, sample_metadata = NULL)
}
\arguments{
\item{counts}{a matrix of data that contains the counts where rows are each paired_guide target and columns are each sample}

\item{pg_metadata}{metadata associated with the pgRNA constructs that correspond to the rows of the counts data}

\item{sample_metadata}{metadata associated with the samples of the dataset that correspond to the columns of the counts data}
}
\description{
This function allows people to have their data ready to be processed by gimap
}
\examples{
\dontrun{
data <- example_data() \%>\%
  dplyr::select(c("Day00_RepA", "Day05_RepA", "Day22_RepA", "Day22_RepB", "Day22_RepC")) \%>\%
  as.matrix()

counts_data <- setup_data(data)
}
}
