% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{example_data}
\alias{example_data}
\title{Returns file paths to example data for package}
\usage{
example_data(which_data)
}
\arguments{
\item{which_data}{options are "bam" or "fastq"; specifies which example dataset should be returned}
}
\description{
This function loads and returns file paths to example data for the packagae. Which dataset is returned must be specified
}
\examples{
\dontrun{

bam_files <- example_data("bam")
fastq_files <- example_data("fastq")
reference_files <- example_data("ref")
}
}
